Version 2.0
Applied patch from Ingo Karkat. Thanks.- ENH: Added a config setting g:calendar_datetime. This allows to remove thedisplay of the current date/time in the title (I don't like it), and offers thewindow's statusline as an alternative.- BUG: The checks for g:calendar_action and g:calendar_sign should also checkfor an existing, but empty variable. Otherwise, it is not possible to disable itin a .vimrc, because the sourcing of calendar.vim would initialize the variables.- BUG: In s:CalendarDoAction(), the check for g:calendar_action must come afterthe "navi" handling; otherwise "navi" breaks if g:calendar_action is disabled(see above).- ENH: The :set wrapscan is ugly, because it is a global setting. The search()commands already pass the 'w' flag, so the only remaining issue were thesearches via /. I modified the 'G0/...' search to 'G$?' (i.e. backward from endof buffer), so that 'wrapscan' isn't required any more. (Even better would be touse search() for these jumps, too.) With this, I can also completely get rid ofthe autocmds in case one does not want the date/time in the title, neither (seeabove).- Using :setlocal buftype=nofile instead of =nowrite; this is more suitable forthis kind of scratch buffer, and avoids that the path may be shown in the title/ statusline (depending on the customization).- BUG: Replaced :setlocal nowrap with 'wrap'. Without this, the 'sidescrolloff'setting may cause the left side of the calendar to disappear if the lastinserted element is near the right window border. 'wrap' shouldn't matter,anyway, and 'sidescrolloff' is a global setting, unfortunately.Try :set sidescrolloff=3 in combination with :let g:calendar_navi = 'bottom' toreproduce.- BUG: The :normal i... for the navi rendering causes a trailing space after the"Next>" button. This way, I cannot quickly type "G$<CR>" to activate the nextbutton. Now using :normal a... to append. This causes the entire navi to shiftone character to the right, but you could fix this by decreasing the columncounter.- ENH: Use :stopinsert in case we've gotten here from insert mode (via<C-O>:Calendar<CR>)...- Using :wincmd w instead of :normal <C-w>; it's simpler. (And you should alwaysuse :normal! (with a bang) to avoid interference with user remappings!)- ENH: I noticed that <left> and <s-left> do the same thing, but in differentways (one uses the navi and the latter is called by the former). I dropped theunintuitive shift mappings and instead do the jumps consistently in the navi,using the b:Calendar... variables instead of that ugly maparg() stuff.- ENH: I noticed that <left> and <up> do the same thing. I changed <up>/<down>to move an entire year, so one can quickly let the years pass...- ENH: the 'q' mapping now returns to the previous window, not the first one.